Use DateTime.Now for DeletedDate in soft deletions
Updated the `DeletedDate` property in `ISoftDeletable` entities
to use `DateTime.Now` instead of `DateTime.UtcNow` in the
`Delete` method of `BaseRepository<T>`.
This change affects the timestamp used for soft deletion, switching from UTC to local system time.
Consider potential impacts on systems relying on UTC for auditing or synchronization.